Find an existing lisp binary for addition to the .emacs file.
Previously, the emacs installer assumed the existence of an mlisp.exe
and mlisp.dxl in whichever Allegro CL installation it found. For
current Express Editions, this is not true.

Now, if an Allegro CL installation is found, it checks for mlisp.exe,
alisp.exe, then Allegro-Express.exe and inserts the first one it finds
into the generated .emacs file. As before, if no lisp installation is
found, then no .emacs file is generated.